requirements.txt
numpy>=1.24
Pillow>=10.0
PySide6>=6.6
numba>=0.58
